Refactor external bq logging to provide reusable controller
